+++ api/lang/es/auth.php
@@ -6,4 +6,9 @@ return [
     'password' => 'La contraseña introducida es incorrecta.',
     'throttle' => 'Demasiados intentos. Vuelve a intentarlo en :seconds segundos.',
 
+    'verification' => [
+        'already_verified' => 'Tu email ya está verificado.',
+        'resent' => 'Te hemos enviado un nuevo enlace de verificación.',
+    ],
+
 ];